+++ src/Aws/DirectoryService/Exception/DirectoryServiceException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Aws\DirectoryService\Exception;
+
+use Aws\Common\Exception\ServiceResponseException;
+
+/**
+ * Exception thrown by the DirectoryService client.
+ */
+class DirectoryServiceException extends ServiceResponseException {}